refactor & add documentation
