+++ client/src/context/RoomsContext.tsx
@@ -1,6 +1,4 @@
-import { useContext } from "react";
-import { useState } from "react";
-import { createContext } from "react";
+import { FC, ReactNode, useState, useContext, createContext } from "react";
 
 export interface RoomProps {
   roomId: string;
@@ -8,26 +6,27 @@ export interface RoomProps {
   members: string[];
 }
 
-interface RoomsContext {
+interface RoomsContextInterface {
   rooms?: RoomProps[];
   getRooms?: (rooms: RoomProps[]) => void;
   addNewRoom?: (room: RoomProps) => void;
 }
 
-export const RoomsContext = createContext<RoomsContext>({});
+export const RoomsContext = createContext<RoomsContextInterface>({});
 
-export const RoomsContextProvider: React.FC<React.ReactNode> = ({
+export const RoomsContextProvider: FC<ReactNode> = ({
   children,
 }) => {
   const [rooms, setRooms] = useState<RoomProps[]>([]);
-  //get rooms
+
   const getRooms = (rooms: RoomProps[]): void => {
     setRooms(rooms);
   };
-  //add new room
+
   const addNewRoom = (room: RoomProps): void => {
     setRooms([...rooms, room]);
   };
+
   return (
     <RoomsContext.Provider value={{ rooms, getRooms, addNewRoom }}>
       {children}
